* move attribute `isScrolling`,`direction` to `ChipGroupStyle` * remove deprecated attribute * fix issue
